+++ packages/website/components/ExampleHeader/index.tsx
@@ -1,0 +1,20 @@
+"use client";
+
+import startCase from "lodash/startCase";
+import { usePathname, useSelectedLayoutSegments } from "next/navigation";
+
+const ExampleHeader = () => {
+    const segments = usePathname()
+        .replace("examples", "example")
+        .split("/")
+        .slice(2)
+        .map(startCase);
+
+    return (
+        <h1>
+            {segments[0]}: {segments[2]} ({segments[1]})
+        </h1>
+    );
+};
+
+export default ExampleHeader;