+++ packages/polyfills/scrollend-polyfill/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @af-utils/scrollend-polyfill
 
+## 0.0.8
+
+### Patch Changes
+
+-   reduced number of dependencies + upgraded website
+
 ## 0.0.7
 
 ### Patch Changes